# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Mentora",
    "contact": {},
    "version": "1.0"
  },
  "servers": [
    {
      "url": "https://mentora-mem.herokuapp.com/",
      "variables": {}
    }
  ],
  "paths": {
...
    "/student": {
      "get": {
        "tags": ["Student"],
        "summary": "ListStudent",
        "operationId": "ListStudent",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      },
      "post": {
        "tags": ["Student"],
        "summary": "CreateStudent",
        "operationId": "CreateStudent",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/CreateStudentRequest"
                  },
                  {
                    "example": {
                      "name": "Example Example",
                      "cpf": "12354759513",
                      "phone": "+5562999626269"
                    }
                  }
                ]
              },
              "example": {
                "name": "Example Example",
                "cpf": "12354759513",
                "phone": "+5562999626269"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false,
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "put": {
        "tags": ["Student"],
        "summary": "UpdateStudent",
        "operationId": "UpdateStudent",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false,
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "tags": ["Student"],
        "summary": "DeleteStudent",
        "operationId": "DeleteStudent",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false,
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateStudentRequest": {
        "title": "CreateStudentRequest",
        "required": ["name", "cpf", "phone"],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "cpf": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          }
        },
        "example": {
          "name": "Example Example",
          "cpf": "12354759513",
          "phone": "+5562999626269"
        }
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "security": [],
...
}